Thay doi database : Book 1 -> N Category | Book N-N Tags -> book_tags Thay doi components vue tuong ung Them input tags va input author Thay doi code : Them xoas sua book , category, author, publisher

+++ app/Http/Controllers/Product/CategoryController.php
@@ -26,15 +26,15 @@ class CategoryController extends Controller
 
         view()->share('top_list_category', $top_items);
     }
+    public function index()
+    {
+        $category = MainModel::all()->toFlatTree()->pluck('name');
+        return $category;
+    }
     public function show(Request $request)
     {
-        $data = MainModel::where('cat_id', $request->id)->with('childrent.books')->with('books')->get();
+        $data = MainModel::where('id', $request->id)->get();
         $items = [];
-        foreach ($data as $key => $value) {
-            foreach ($value->books as $key => $value) {
-                $items[] = $value;
-            }
-        }
         if (count($items) > 0) {
             return  response()->json($items, 200, ["Content-type: application/json", "charset=utf-8"]);
         } else {
@@ -45,19 +45,18 @@ class CategoryController extends Controller
     {
         try {
             DB::beginTransaction();
-            $data = new MainModel();
-            $data->cat_id = $request->cat_id;
-            $data->cat_name = $request->cat_name;
-            $data->type_id = $request->type == null ? $request->new_type : $request->type;
-            $data->description  = $request->content;
-            $data->created_by = Auth::user()->user_name;
-            $data->save();
+            $data = MainModel::create([
+                'name' => $request->name,
+                'parent_id' => $request->cat_group ==  null ? null : $request->cat_group,
+                'description' => $request->content == null ? "This category has been added by" . Auth::user()->user_name : $request->content,
+                'created_by' => Auth::user()->user_name
+            ]);
             DB::commit();
-            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Add ' . $request->cat_name . ' Successfully !! </div>');
+            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Add ' . $request->name . ' Successfully !! </div>');
             return \redirect()->back();
         } catch (Exception $e) {
             DB::rollBack();
-            $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Add ' . $request->cat_name . 'failed with error ' . $e->getMessage() . ' Try Again !!</div>');
+            $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Add ' . $request->name . 'failed with error ' . $e->getMessage() . ' Try Again !!</div>');
             return \redirect()->back();
         }
     }
@@ -65,16 +64,16 @@ class CategoryController extends Controller
     {
         try {
             DB::beginTransaction();
-            $data = MainModel::withTrashed()->where('cat_id', $request->cat_id)->first();
-            $data->cat_name = $request->cat_name;
+            $data = MainModel::withTrashed()->where('id', $request->id)->first();
+            $data->name = $request->name;
             $data->description = $request->content;
-            $data->type_id = $request->type == null ? $request->new_type : $request->type;
+            $data->parent_id = $request->cat_group == null ? null : $request->cat_group;
             $data->modified_by = Auth::user()->user_name;
             if ($data->isDirty() == false) {
                 $request->session()->flash(
                     'infor_mess',
                     '<div class="alert alert-primary" style="text-align: center;font-size: x-large;font-family: fangsong;"> 
-                        You have yet to update the information for "' . $data->getOriginal('cat_name') . '"! 
+                        You have yet to update the information for "' . $data->getOriginal('name') . '"! 
                         </br>
                         Please check it or return back if you dont want to update this category ! 
                     </div>'
@@ -82,12 +81,18 @@ class CategoryController extends Controller
                 return redirect()->back();
             }
             $data->save();
+            $check_tree = MainModel::isBroken();
+            if ($check_tree == true) {
+                $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> The Tree Category will be broken ! Check it and try Again !! </div>');
+                DB::rollBack();
+                return \redirect()->back();
+            }
             DB::commit();
-            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Edit ' . $request->cat_name . ' Successfully !! </div>');
+            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Edit ' . $request->name . ' Successfully !! </div>');
             return \redirect()->route('admin.category');
         } catch (Exception $e) {
             DB::rollBack();
-            $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Edit ' . $request->cat_name . 'Fail,Try Again !! </div>');
+            $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Edit ' . $request->name . 'Fail,Try Again !! </div>');
             return \redirect()->back();
         }
     }
@@ -95,20 +100,26 @@ class CategoryController extends Controller
     {
         try {
             DB::beginTransaction();
-            $data = MainModel::withTrashed()->where('cat_id', $request->cat_id)->first();
-            if ($data->books()->exists() == false) {
+            $data = MainModel::withTrashed()->where('id', $request->id)->first();
+            if ($data->books()->exists() == false && $data->isLeaf()) {
                 $data->forceDelete();
-                $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Delete  category ' . $request->cat_name . ' Successfully !! </div>');
+                $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Delete  category ' . $request->name . ' Successfully !! </div>');
             } else {
                 $data->deleted_by = Auth::user()->user_name;
                 $data->save();
                 $data->delete();
-                $request->session()->flash('infor_mess', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> The category ' . $request->cat_name . ' has been moved to the "Restore List" category!! </div>');
+                $request->session()->flash('infor_mess', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> The category ' . $request->name . ' has been moved to the "Restore List" category!! </div>');
+            }
+            $check_tree = MainModel::isBroken();
+            if ($check_tree == true) {
+                $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> The Tree Category will be broken ! Check it and try Again !! </div>');
+                DB::rollBack();
+                return \redirect()->back();
             }
             DB::commit();
             return redirect()->back();
         } catch (Exception $e) {
-            $request->session()->flash('info_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Delete  category ' . $request->cat_name . 'Failed with error ' . $e->getMessage() .  '</div>');
+            $request->session()->flash('info_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Delete  category ' . $request->name . 'Failed with error ' . $e->getMessage() .  '</div>');
             DB::rollBack();
             return redirect()->back();
         }
@@ -117,11 +128,19 @@ class CategoryController extends Controller
     {
         try {
             DB::beginTransaction();
-            $data = MainModel::withTrashed()->where('cat_id', $request->cat_id)->first();
+            $category = MainModel::onlyTrashed()->get();
+            $data = MainModel::withTrashed()->where('id', $request->id)->first();
+            foreach ($category as $key => $value) {
+                if ($data->isChildOf($value)) {
+                    $request->session()->flash('infor_warning', '<div class="alert alert-danger" style="text-align: center;font-size: x-large;font-family: fangsong;"> Please restore the parent category of this category first and try again !! The parent of this category is </br> "' . $value->name . '"</div>');
+                    DB::rollBack();
+                    return \redirect()->back();
+                }
+            }
             $data->deleted_by = null;
             $data->save();
             $data->restore();
-            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Restore category ' . $data->cat_name . ' Successfully !! </div>');
+            $request->session()->flash('infor_success', '<div class="alert alert-success" style="text-align: center;font-size: x-large;font-family: fangsong;"> Restore category ' . $data->name . ' Successfully !! </div>');
             DB::commit();
             return redirect()->back();
         } catch (\Throwable $th) {
